move cfg in another directory
