+++ ydb/core/ymq/actor/actor.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "defs.h"
+#include <ydb/core/ymq/actor/cfg/defs.h>
 
 #include <ydb/core/protos/msgbus.pb.h>
 #include <ydb/library/actors/core/actor.h>
