+++ ydb/core/ymq/actor/attributes_md5.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "defs.h"
+#include <ydb/core/ymq/actor/cfg/defs.h>
 #include <ydb/core/protos/sqs.pb.h>
 
 namespace NKikimr::NSQS {
