+++ ydb/core/ymq/actor/auth_multi_factory.cpp
@@ -1,7 +1,7 @@
 #include "auth_multi_factory.h"
 
 #include <ydb/core/base/ticket_parser.h>
-#include <ydb/core/ymq/actor/cfg.h>
+#include <ydb/core/ymq/actor/cfg/cfg.h>
 #include <ydb/core/ymq/actor/error.h>
 #include <ydb/core/ymq/actor/proxy_actor.h>
 #include <ydb/core/ymq/actor/serviceid.h>
