+++ ydb/core/ymq/actor/cfg/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+SRCS(
+    cfg.cpp
+)
+
+PEERDIR(
+    ydb/core/base
+    ydb/core/protos
+    ydb/library/actors/core
+)
+
+END()
