+++ ydb/core/ymq/actor/garbage_collector.cpp
@@ -1,6 +1,6 @@
 #include "garbage_collector.h"
 
-#include "cfg.h"
+#include <ydb/core/ymq/actor/cfg/cfg.h>
 #include "log.h"
 #include "events.h"
 #include "executor.h"
