+++ ydb/core/ymq/actor/local_rate_limiter_allocator.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "defs.h"
+#include <ydb/core/ymq/actor/cfg/defs.h>
 #include <util/system/types.h>
 
 namespace NKikimr::NSQS {
