+++ ydb/core/ymq/actor/message_delay_stats.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "defs.h"
+#include <ydb/core/ymq/actor/cfg/defs.h>
 
 #include <util/datetime/base.h>
