+++ ydb/core/ymq/actor/metering.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "defs.h"
+#include <ydb/core/ymq/actor/cfg/defs.h>
 
 #include <ydb/core/ymq/base/action.h>
 #include <ydb/core/util/address_classifier.h>
