+++ ydb/core/ymq/actor/purge.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "defs.h"
+#include <ydb/core/ymq/actor/cfg/defs.h>
 #include "events.h"
 
 #include <ydb/library/actors/core/actor_bootstrapped.h>
