+++ ydb/core/ymq/actor/retention.cpp
@@ -1,5 +1,5 @@
 #include "retention.h"
-#include "cfg.h"
+#include <ydb/core/ymq/actor/cfg/cfg.h>
 #include "log.h"
 #include "executor.h"
