+++ ydb/core/ymq/actor/serviceid.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "defs.h"
+#include <ydb/core/ymq/actor/cfg/defs.h>
 
 #include <ydb/library/actors/core/actor.h>
