+++ ydb/core/ymq/actor/user_settings_names.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "defs.h"
+#include <ydb/core/ymq/actor/cfg/defs.h>
 
 #include <util/generic/string.h>
