+++ ydb/core/ymq/actor/user_settings_reader.cpp
@@ -1,5 +1,5 @@
 #include "user_settings_reader.h"
-#include "cfg.h"
+#include <ydb/core/ymq/actor/cfg/cfg.h>
 #include "executor.h"
 #include "events.h"
